Chain sign-in onto sign-up, wake up heroku
Add sign in to promise chain on successful sign up. Move jQuery button
click of sign in to the event method so that it isn't also triggered on
sign up.
Closes #8

Add wakeUpHeroku AJAX req to authAPI to make a get request to the API
on document load, so hopefully heroku's dynos will be woken up before
the user tries to sign in, preventing the annoying load time.
closes #7